+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.6] - 2025-12-17
+
+### 🐛 Bug Fixes
+- CSV Export: Fixed start date initialization to use actual earliest date from data store instead of sampled data.
+- CSV Export: Removed non-functional "Preview" button from export modal.
+
+### ✨ Features
+- CSV Export: Added "Export All" option in export modal to quickly export all historical data without selecting a date range.
+
 ## [1.4.5] - 2025-12-17
 
 ### 🛠 Changes